Binary banana dataset

Dataset containing two attributes as well as a class one, that, if plotted, represent a banana shape

banana

banana_orig

Format

At1

First attribute.

At2

Second attribute.

Class

Two possible classes: positive (banana shape), negative (surrounding of the banana).

Source

KEEL Repository.

Shape

banana: A data frame with 2640 instances, 264 of which belong to positive class, and 3 variables

banana_orig: A data frame with 5300 instances, 2376 of which belong to positive class, and 3 variables: