Fitering of oversampled data based on non-cooperative game theory

Filters oversampled examples from a binary class dataset using game theory to find out if keeping an example is worthy enough.

neater(dataset, newSamples, k = 3, iterations = 100, smoothFactor = 1,
  classAttr = "Class")

Arguments

dataset
The original data.frame. All columns, except classAttr one, have to be numeric or coercible to numeric.

newSamples
A data.frame containing the samples to be filtered. Must have the same structure as dataset.

k
Integer. Number of nearest neighbours to use in KNN algorithm to rule out samples. By default, 3.

iterations
Integer. Number of iterations for the algorithm. By default, 100.

smoothFactor
A positive numeric. By default, 1.

classAttr
character. Indicates the class attribute from dataset and newSamples. Must exist in them.

Value

Filtered samples as a data.frame with same structure as newSamples.

Details

Uses game theory and Nash equilibriums to calculate the minority examples probability of trully belonging to the minority class. It discards examples which at the final stage of the algorithm have more probability of being a majority example than a minority one.

References

Almogahed, B.A.; Kakadiaris, I.A. Neater: Filtering of Over-Sampled Data Using Non-Cooperative Game Theory. Soft Computing 19 (2014), Nr. 11, p. 3301–3322.

Examples

data(iris0)

newSamples <- smotefamily::SMOTE(iris0[,-5], iris0[,5])$syn_data
# SMOTE overrides Class attr turning it into class
# and dataset must have same class attribute as newSamples
names(newSamples) <- c(names(newSamples)[-5], "Class")

neater(iris0, newSamples, k = 5, iterations = 100,
       smoothFactor = 1, classAttr = "Class")

#> [1] "0 samples filtered by NEATER"

#>    SepalLength SepalWidth PetalLength PetalWidth    Class
#> 1     5.277446   3.500000    1.448369  0.2000000 positive
#> 2     4.822336   3.400000    1.588832  0.2000000 positive
#> 3     4.780010   3.119990    1.600000  0.2000000 positive
#> 4     4.494730   2.336891    1.300000  0.2947299 positive
#> 5     4.660417   3.279165    1.520835  0.2395826 positive
#> 6     4.353876   3.000000    1.207752  0.1538760 positive
#> 7     5.507364   4.207364    1.403682  0.2073636 positive
#> 8     5.100000   3.500000    1.400000  0.2161409 positive
#> 9     4.594136   3.194136    1.397068  0.2000000 positive
#> 10    5.011491   3.622982    1.411491  0.2114909 positive
#> 11    5.496510   4.189530    1.410470  0.2069799 positive
#> 12    5.271106   3.400000    1.435553  0.2711060 positive
#> 13    4.600000   3.474135    1.125865  0.2000000 positive
#> 14    5.045966   3.345966    1.445966  0.2000000 positive
#> 15    5.196883   3.700000    1.500000  0.3031171 positive
#> 16    4.614348   3.371305    1.385652  0.2856523 positive
#> 17    5.799959   3.999990    1.200010  0.2000207 positive
#> 18    4.805478   3.000000    1.400000  0.1054780 positive
#> 19    5.111907   3.488093    1.400000  0.2000000 positive
#> 20    4.899598   3.099598    1.499598  0.1000000 positive
#> 21    4.556153   3.200000    1.378076  0.2000000 positive
#> 22    4.927230   3.000000    1.454460  0.2000000 positive
#> 23    5.038198   3.438198    1.523604  0.3618018 positive
#> 24    5.000000   3.415613    1.515613  0.2624519 positive
#> 25    4.411524   3.200000    1.300000  0.2000000 positive
#> 26    5.400000   3.750145    1.550145  0.2501455 positive
#> 27    5.120433   3.479567    1.400000  0.2795671 positive
#> 28    5.100000   3.622963    1.461482  0.3614817 positive
#> 29    5.178166   4.034498    1.521834  0.1218338 positive
#> 30    5.590886   3.836371    1.700000  0.3363712 positive
#> 31    5.115584   3.500000    1.415584  0.2000000 positive
#> 32    4.800000   3.400000    1.787060  0.2000000 positive
#> 33    4.649785   3.149785    1.549785  0.2000000 positive
#> 34    5.000000   3.308718    1.408718  0.2000000 positive
#> 35    4.800000   3.000000    1.400000  0.2884264 positive
#> 36    5.011519   3.500000    1.311519  0.3000000 positive
#> 37    5.241312   3.847104    1.805792  0.4000000 positive
#> 38    4.850961   3.100000    1.549039  0.1490393 positive
#> 39    5.305654   3.700000    1.500000  0.2000000 positive
#> 40    5.049774   3.350226    1.599547  0.3493207 positive
#> 41    4.825843   3.100000    1.574157  0.1741574 positive
#> 42    5.109461   3.803154    1.493693  0.3031537 positive
#> 43    5.000000   3.438875    1.279625  0.2796251 positive
#> 44    5.088474   3.511526    1.400000  0.2000000 positive
#> 45    4.654847   3.072576    1.472576  0.2274236 positive
#> 46    5.295314   3.452343    1.595314  0.2000000 positive
#> 47    5.009857   3.409857    1.490143  0.2000000 positive
#> 48    4.400000   3.034920    1.355027  0.2000000 positive
#> 49    5.229901   3.735049    1.535049  0.2000000 positive
#> 50    4.900000   3.100000    1.500000  0.1000000 positive