Random Walk Oversampling

Generate synthetic minority examples for a dataset without modifying their variance or mean.

rwo(dataset, numInstances, classAttr = "Class")

Arguments

dataset
data.frame to treat

numInstances
Integer. Number of new minority examples to generate.

classAttr
String. Indicates the class attribute from dataset. Must exist in it.

Value

new samples, a data.frame with the same structure as dataset, containing the synthetic examples generated

Details

Generates numInstances new minority examples for dataset, adding to the each column of the j-th example its variance scalated by the inverse of the number of minority examples and a factor following a N(0,1) distribution which depends on the example.

Examples

data(iris0)

newSamples <- rwo(iris0, numInstances = 100, classAttr = "Class")