Wrapper for Rapidy Converging Gibbs algorithm.

Generates synthetic minority examples by approximating their probability distribution until sensitivity of wrapper over validation cannot be further improved. Works only on discrete numeric datasets

wracog(train, validation, wrapper, classAttr = "Class", slideWin = 10,
  threshold = 0.02)

Arguments

train
data.frame. A initial dataset to generate first model. All columns, except classAttr one, have to be numeric

validation
data.frame. A dataset to compare results of consecutive classifiers. Must have the same structure of train.

wrapper
An S3 object. There must be a method trainWrapper implemented for the class of the object, and a predict method implemented for the class of the model returned by trainWrapper.

classAttr
String. Indicates the class attribute from dataset. Must exist in it.

slideWin
Number of last sensitivities to take into account to meet the stopping criteria. By default, 10.

threshold
Threshold that the last slideWin sensitivities mean should reach. By default, 0.02.

Value

new samples, a data.frame with the same structure as dataset, containing the synthetic examples generated

Details

Until the last slideWin executions of wrapper over validation dataset reach a mean sensitivity lower than threshold, the algorithm keeps generating samples using Gibbs Sampler, and adding misclassified samples with respect to a model generated by a former train, to the train dataset. Iniial model is built on initial train.

Examples

data(haberman)
myWrapper <- structure(list(), class="C50Wrapper")
trainWrapper.C50Wrapper <- function(wrapper, train, trainClass){
  C50::C5.0(train, trainClass)
}

trainFold <- sample(1:nrow(haberman), nrow(haberman)/2, FALSE)
wracog(haberman[trainFold, ], haberman[-trainFold, ],
       myWrapper, classAttr = "Class")

#> Error in wracog(haberman[trainFold, ], haberman[-trainFold, ], myWrapper,     classAttr = "Class"): There doesn't exist a method trainWrapper.C50Wrapper